Added more tests to valid ISBN10 codes -> fails

+++ src/main/java/com/lvl/isbnTools/ISBNValidator.java
@@ -2,7 +2,7 @@ package com.lvl.isbnTools;
 
 public class ISBNValidator {
 
-	public boolean isValidISBN(int isbn) {
+	public boolean isValidISBN(long isbn) {
 		if (isbn == 1719587213)
 			return true;
 		return false;
